# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SHOP.CO API Documentation",
        "description": "Tài liệu tích hợp API cho hệ thống E-commerce",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Dynamic Server (Follow file .env)"
        },
        {
            "url": "https://api.quocnva09.me/api",
            "description": "Production Server (Cloudflare SSL)"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Local Server (Dự phòng)"
        }
    ],
    "paths": {
...
        "/api/admin/exports": {
            "get": {
                "tags": [
                    "Export Module - Admin"
                ],
                "summary": "List the current admin's export history",
                "operationId": "8194f158161a9b397dc3240b200ce41c",
                "responses": {
                    "200": {
                        "description": "Export history retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ExportResource"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginatedMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Export Module - Admin"
                ],
                "summary": "Trigger a new product export job",
                "operationId": "49e351f2adc5b90ff896bb986156bf9f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExportRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Export processing started successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Export processing started successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ExportResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/exports/{id}": {
            "get": {
                "tags": [
                    "Export Module - Admin"
                ],
                "summary": "Get a single export job record",
                "operationId": "09e5eb0b6c7c0634d34118d22fa0d44f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Export record retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ExportResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/exports/{id}/download": {
            "get": {
                "tags": [
                    "Export Module - Admin"
                ],
                "summary": "Download a completed export file",
                "operationId": "c4c3100a099bbf5adde90c678e374f24",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File downloaded successfully",
                        "content": {
                            "application/octet-stream": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Export not ready or file not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Export file is not ready yet"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "PaginatedMeta": {
                "title": "Pagination Meta",
                "description": "Các thông số toán học để vẽ giao diện phân trang",
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "total": {
                        "type": "integer",
                        "example": 150
                    }
                },
                "type": "object"
            },
...
            "ExportRequest": {
                "title": "Export Request",
                "description": "Payload to trigger a new product export job",
                "properties": {
                    "format": {
                        "type": "string",
                        "example": "csv",
                        "enum": [
                            "csv",
                            "xlsx"
                        ]
                    },
                    "search": {
                        "type": "string",
                        "example": "shirt",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "nullable": true,
                        "enum": [
                            "pending",
                            "processing",
                            "completed",
                            "failed"
                        ]
                    }
                },
                "type": "object"
            },
...
            "ExportResource": {
                "title": "Export Resource",
                "description": "An export job record tracking the status and download path of a generated file",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "type": "string",
                        "example": "products"
                    },
                    "format": {
                        "type": "string",
                        "example": "csv",
                        "enum": [
                            "csv",
                            "xlsx"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "example": "completed",
                        "enum": [
                            "pending",
                            "processing",
                            "completed",
                            "failed"
                        ]
                    },
                    "file_path": {
                        "type": "string",
                        "example": "exports/products_2024-01-15.csv",
                        "nullable": true
                    },
                    "error_message": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:30:00+07:00",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Nhập Sanctum Personal Access Token của bạn vào đây",
                "scheme": "bearer"
            }
        }
    },
...
}